==> notebooks/01-jimmy-wrangler-trump-tweets.ipynb <==
# Jimmy Wrangler, Data Explorer

Dalton Hahn (2762306)

## realDonaldTrump Poll Tweets
https://github.com/fivethirtyeight/data/blob/master/trump-twitter/realDonaldTrump_poll_tweets.csv

## Exploratory Data Analysis
My idea is to examine the three datasets to see if there is any correlation or connection between President Trump's tweets and the lawsuits he and his company have been involved in.  Additionally, I would like to see if there is any correlation or connection between his tweets and his overall approval rating.

import pandas as pd
import numpy as np
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt
import math
from statistics import mean, stdev

## Read in the Data

df = pd.read_csv("../data/external/realDonaldTrump_poll_tweets.csv")

df.head(1)

print(df['created_at'][0])

## Splitting Timestamp
Split the timestame into four components
- Month
- Day
- Year
- Time

months = []
days = []
years = []
times = []

for row in df['created_at']:
    month = row.split('/')[0]
    months.append(month)
    #print(month)
    day = row.split('/')[1]
    days.append(day)
    #print(day)
    year = row.split('/')[2].split(' ')[0]
    years.append(year)
    #print(year)
    time = row.split('/')[2].split(' ')[1]
    times.append(time)
    #print(time)

#Check that the splits I've created are the same size as the original data frame
print(len(days))
print(len(df['created_at']))

## Add the newly created columns to the original dataframe

df.insert(len(df.columns), 'Month', months, True)
df.insert(len(df.columns), 'Day', days, True)
df.insert(len(df.columns), 'Year', years, True)
df.insert(len(df.columns), 'Time', time, True)

df.head()

## Remove arbitrary ID value that came in with data

df = df.drop(axis=1, columns='id')
df.head()

## Save out edited data into data/processed for use in analysis notebook

df.to_csv("../data/processed/tweets_processed.csv")
