==> notebooks/03-jimmy-wrangler-trump-lawsuits.ipynb <==
# Jimmy Wrangler, Data Explorer

Dalton Hahn (2762306)

## Donald Trump Lawsuits
https://raw.githubusercontent.com/fivethirtyeight/data/master/trump-lawsuits/trump-lawsuits.csv

## Exploratory Data Analysis
My idea is to examine the three datasets to see if there is any correlation or connection between President Trump's tweets and the lawsuits he and his company have been involved in.  Additionally, I would like to see if there is any correlation or connection between his tweets and his overall approval rating.

import pandas as pd
import numpy as np
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt
import math
from statistics import mean, stdev

## Read in the Data

df = pd.read_csv("../data/external/trump-lawsuits.csv")

df.head(1)

## I only care about the date that a lawsuit was filed, so I will remove all other columns.  I'll keep the column for issue to have some context as to what the lawsuit was regarding.

# Dropping columns from approval_topline
df = df.drop(axis=1, columns=['docketNumber', 'caseName', 'plaintiff', 'defendant', 
                              'currentLocation', 'previousLocation', 'jurisdiction', 'judge', 'nature',
                             'TrumpCategory', 'capacity', 'type', 'docketOrig', 'status'])
df.head()

## Will break down the date field into component parts for easier parsing/comparison in the analysis notebook

months = []
days = []
years = []

for row in df['dateFiled']:
    month = row.split('-')[1]
    months.append(month)
    #print(month)
    day = row.split('-')[2]
    days.append(day)
    #print(day)
    year = row.split('-')[0]
    years.append(year)
    #print(year)

#Check that the splits I've created are the same size as the original data frame
print(len(days))
print(len(df['dateFiled']))

df.insert(len(df.columns), 'Month', months, True)
df.insert(len(df.columns), 'Day', days, True)
df.insert(len(df.columns), 'Year', years, True)
df.head()

### Save out edited data into data/processed for use in analysis notebook

df.to_csv("../data/processed/topline.csv")
