==> notebooks/04-jimmy-wrangler-trump-analysis.ipynb <==
# Jimmy Wrangler, Data Explorer

Dalton Hahn (2762306)

## Donald Trump Combined Analysis

## Exploratory Data Analysis
My idea is to examine the three datasets to see if there is any correlation or connection between President Trump's tweets and the lawsuits he and his company have been involved in.  Additionally, I would like to see if there is any correlation or connection between his tweets and his overall approval rating.

import pandas as pd
import numpy as np
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt
import math
from statistics import mean, stdev

## Read in the Data

df_tweet = pd.read_csv("../data/processed/tweets_processed.csv")
df_tweet.head()

df_poll = pd.read_csv("../data/processed/polling.csv")
df_poll.head()

df_top = pd.read_csv("../data/processed/topline.csv")
df_top.head()

df_lawsuits = pd.read_csv("../data/processed/lawsuits.csv")
df_lawsuits.head()

## First checking to see how closely his tweets correlate with the lawsuits brought against him

t_to_suit = {}


for index, t_row in df_tweet.iterrows():
    for index, l_row in df_lawsuits.iterrows():
        if t_row['Year'] == l_row['Year'] and t_row['Month'] == l_row['Month']:
            t_to_suit[t_row['text']] = l_row['issue']
            print(t_row['Month'], t_row['Year'])

# for key,val in t_to_suit.items():
#     print(key, "=>", val)
#     print()
    
df_t_to_suit = pd.DataFrame.from_dict(t_to_suit, orient='index')
df_t_to_suit.head()

## Does not appear that any of the tweets that he made in months were a lawsuit was filed against him have any correlation to one another

- Looking into polling ratings for the times of these 5 tweets (8/2016 and 3/2016)

for index, row in df_poll.iterrows():
    if row['M_Year'] == 2016 and row['M_Month'] == 3:
        print(row)
    elif row['M_Year'] == 2016 and row['M_Month'] == 8:
        print(row)
    elif row['C_Year'] == 2016 and row['C_Month'] == 3:
        print(row)
    elif row['C_Year'] == 2016 and row['C_Month'] == 8:
        print(row)


## No data corresponding for the times of the tweets out of the Approval Polling dataset, checking the Approval Topline dataset

for index, row in df_top.iterrows():
    if row['M_Year'] == 2016 and row['M_Month'] == 3:
        print(row)
    elif row['M_Year'] == 2016 and row['M_Month'] == 8:
        print(row)

## No data corresponding for the times of the tweets out of the Approval Topline dataset either, widening the search to just the year 2016

for index, row in df_top.iterrows():
    if row['M_Year'] == 2016:
        print(row)
        
for index, row in df_poll.iterrows():
    if row['M_Year'] == 2016 or row['C_Year'] == 2016:
        print(row)

## Conclusions
- With the datasets provided, the only correlations that were able to be made about President Trump's lawsuits, approval ratings, and tweets were between lawsuits and tweets.
- Time was considered as the correlating factor and was examined at the month-level granularity
- Taking this knowledge and applying it between the tweet data and the lawsuit data, 5 elements were found to be in correlation
- Extending this correlation further to approval ratings however was unfortunately not possible due to a lack of data in the approval datasets from this time
- In addition to attempting to extend the correlation, the granularity that was examined was also widened to be year-level granularity which unfortunately, still produced no results in the datasets
